# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisResponse": {
        "additionalProperties": false,
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/AnalysisResult"
          }
        },
        "required": [
          "analysis"
        ],
        "title": "AnalysisResponse",
        "type": "object"
      },
      "AnalysisResult": {
        "additionalProperties": false,
        "description": "Complete analysis output for an uploaded dataset.",
        "properties": {
          "analysis_id": {
            "title": "Analysis Id",
            "type": "string"
          },
          "anomalies": {
            "items": {
              "$ref": "#/components/schemas/AnomalyFlag"
            },
            "title": "Anomalies",
            "type": "array"
          },
          "behavioral_insights": {
            "items": {
              "$ref": "#/components/schemas/BehavioralInsight"
            },
            "title": "Behavioral Insights",
            "type": "array"
          },
          "benford_chi_squared": {
            "minimum": 0.0,
            "title": "Benford Chi Squared",
            "type": "number"
          },
          "benford_digits": {
            "items": {
              "$ref": "#/components/schemas/BenfordDigitResult"
            },
            "title": "Benford Digits",
            "type": "array"
          },
          "benford_is_suspicious": {
            "default": false,
            "title": "Benford Is Suspicious",
            "type": "boolean"
          },
          "benford_p_value": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Benford P Value",
            "type": "number"
          },
          "completed_at": {
            "format": "date-time",
            "title": "Completed At",
            "type": "string"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "graph_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphAnalysisResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "investigation_leads": {
            "items": {
              "$ref": "#/components/schemas/InvestigationLead"
            },
            "title": "Investigation Leads",
            "type": "array"
          },
          "outlier_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Outlier Count",
            "type": "integer"
          },
          "outlier_pct": {
            "default": 0.0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Outlier Pct",
            "type": "number"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "risk_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score",
            "type": "integer"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "total_anomalies": {
            "minimum": 0.0,
            "title": "Total Anomalies",
            "type": "integer"
          },
          "total_transactions": {
            "minimum": 0.0,
            "title": "Total Transactions",
            "type": "integer"
          },
          "velocity_spikes": {
            "items": {
              "$ref": "#/components/schemas/VelocitySpike"
            },
            "title": "Velocity Spikes",
            "type": "array"
          }
        },
        "required": [
          "analysis_id",
          "dataset_id",
          "completed_at",
          "total_transactions",
          "total_anomalies",
          "risk_score",
          "risk_level",
          "benford_chi_squared",
          "benford_p_value"
        ],
        "title": "AnalysisResult",
        "type": "object"
      },
...
      "Body_upload_dataset_api_v1_datasets_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_dataset_api_v1_datasets_upload_post",
        "type": "object"
      },
...
      "CreateCaseFromAnalysisResult": {
        "additionalProperties": false,
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/AnalysisResult"
          },
          "case": {
            "$ref": "#/components/schemas/FraudCase"
          },
          "linked_alerts": {
            "items": {
              "$ref": "#/components/schemas/FraudAlert"
            },
            "title": "Linked Alerts",
            "type": "array"
          }
        },
        "required": [
          "analysis",
          "case",
          "linked_alerts"
        ],
        "title": "CreateCaseFromAnalysisResult",
        "type": "object"
      },
...
      "DatasetListResponse": {
        "additionalProperties": false,
        "properties": {
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DatasetResponse"
            },
            "title": "Datasets",
            "type": "array"
          }
        },
        "required": [
          "datasets"
        ],
        "title": "DatasetListResponse",
        "type": "object"
      },
      "DatasetResponse": {
        "additionalProperties": false,
        "properties": {
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/DatasetStatus"
          },
          "uploaded_at": {
            "title": "Uploaded At",
            "type": "string"
          }
        },
        "required": [
          "dataset_id",
          "name",
          "uploaded_at",
          "row_count",
          "status"
        ],
        "title": "DatasetResponse",
        "type": "object"
      },
      "DatasetStatus": {
        "enum": [
          "uploaded",
          "analyzing",
          "completed",
          "failed"
        ],
        "title": "DatasetStatus",
        "type": "string"
      },
...
      "ExplanationAudience": {
        "enum": [
          "analyst",
          "admin"
        ],
        "title": "ExplanationAudience",
        "type": "string"
      },
...
      "FraudAlert": {
        "additionalProperties": false,
        "description": "An auto-generated or manually created alert that may lead to a case.",
        "properties": {
          "acknowledged_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "alert_id": {
            "title": "Alert Id",
            "type": "string"
          },
          "assigned_analyst_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Analyst Id"
          },
          "assigned_analyst_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Analyst Name"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "linked_case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linked Case Id"
          },
          "narrative": {
            "title": "Narrative",
            "type": "string"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "rule_code": {
            "title": "Rule Code",
            "type": "string"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "severity": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/WorkflowSourceType",
            "default": "scenario"
          },
          "status": {
            "$ref": "#/components/schemas/AlertStatus",
            "default": "new"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "alert_id",
          "source_id",
          "rule_code",
          "title",
          "severity",
          "narrative",
          "created_at"
        ],
        "title": "FraudAlert",
        "type": "object"
      },
      "FraudCase": {
        "additionalProperties": false,
        "description": "A persistent fraud investigation case with lifecycle management.",
        "properties": {
          "alert_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Alert Count",
            "type": "integer"
          },
          "assigned_analyst_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Analyst Id"
          },
          "assigned_analyst_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Analyst Name"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "comment_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Comment Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "disposition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseDisposition"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "$ref": "#/components/schemas/CasePriority",
            "default": "medium"
          },
          "resolution_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Notes"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "risk_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score",
            "type": "integer"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "sla_deadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Deadline"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/WorkflowSourceType",
            "default": "scenario"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus",
            "default": "open"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "source_id",
          "title",
          "risk_score",
          "risk_level",
          "summary",
          "created_at",
          "updated_at"
        ],
        "title": "FraudCase",
        "type": "object"
      },
...
      "GetAnalysisExplanationResult": {
        "additionalProperties": false,
        "properties": {
          "explanation": {
            "$ref": "#/components/schemas/AnalysisExplanation"
          }
        },
        "required": [
          "explanation"
        ],
        "title": "GetAnalysisExplanationResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RiskLevel": {
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "RiskLevel",
        "type": "string"
      },
...
      "TransactionIngestBody": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "default": "api-ingestion",
            "title": "Name",
            "type": "string"
          },
          "transactions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Transactions",
            "type": "array"
          }
        },
        "required": [
          "transactions"
        ],
        "title": "TransactionIngestBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "A fraud triage workspace for uploaded transaction datasets, with persistent alerts and cases, plus reference scenario investigations for validation and rule calibration. Statistical and behavioral analysis remain the source of truth, while the optional copilot layer explains the results in operator-facing language.",
    "title": "Relational Fraud Intelligence",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/datasets": {
      "get": {
        "operationId": "list_datasets_api_v1_datasets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List uploaded datasets",
        "tags": [
          "Datasets"
        ]
      }
    },
    "/api/v1/datasets/ingest": {
      "post": {
        "description": "Accepts a JSON array of transaction records for programmatic ingestion.",
        "operationId": "ingest_transactions_api_v1_datasets_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionIngestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ingest transactions via API",
        "tags": [
          "Datasets"
        ]
      }
    },
    "/api/v1/datasets/upload": {
      "post": {
        "description": "Upload a CSV file with transaction data for fraud analysis. Required columns: transaction_id, account_id, amount, timestamp. Optional: merchant, category, device_fingerprint, ip_country, channel, is_fraud.",
        "operationId": "upload_dataset_api_v1_datasets_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_dataset_api_v1_datasets_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload a transaction CSV",
        "tags": [
          "Datasets"
        ]
      }
    },
    "/api/v1/datasets/{dataset_id}/analysis": {
      "get": {
        "operationId": "get_analysis_results_api_v1_datasets__dataset_id__analysis_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get analysis results for a dataset",
        "tags": [
          "Datasets"
        ]
      }
    },
    "/api/v1/datasets/{dataset_id}/analyze": {
      "post": {
        "description": "Executes Benford's Law analysis, statistical outlier detection, velocity spike detection, round-amount structuring detection, and behavioral relationship inference over accounts, devices, merchants, and geographies.",
        "operationId": "analyze_dataset_api_v1_datasets__dataset_id__analyze_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run fraud analysis on a dataset",
        "tags": [
          "Datasets"
        ]
      }
    },
    "/api/v1/datasets/{dataset_id}/case": {
      "post": {
        "description": "Creates a persistent fraud case from a completed dataset analysis using the analysis-generated investigation leads, and links any open dataset alerts to that case.",
        "operationId": "create_case_from_analysis_api_v1_datasets__dataset_id__case_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCaseFromAnalysisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a case from a dataset analysis",
        "tags": [
          "Datasets"
        ]
      }
    },
    "/api/v1/datasets/{dataset_id}/explanation": {
      "get": {
        "description": "Returns an operator-facing explanation of a completed dataset analysis. Statistical and behavioral scoring remain the source of truth even when the optional Hugging Face explanation provider is active.",
        "operationId": "get_analysis_explanation_api_v1_datasets__dataset_id__explanation_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "audience",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ExplanationAudience",
              "default": "analyst"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAnalysisExplanationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Explain a dataset analysis",
        "tags": [
          "Datasets"
        ]
      }
    },
...
  },
...
}